Form data adapter
