+++ src/collective/volto/formsupport/adapters/configure.zcml
@@ -1,0 +1,10 @@
+<configure
+    xmlns="http://namespaces.zope.org/zope"
+    xmlns:plone="http://namespaces.plone.org/plone"
+    i18n_domain="collective.volto.formsupport"
+    >
+
+  <adapter factory=".FormDataAdapter" />
+
+</configure>
+
