+++ src/collective/volto/formsupport/configure.zcml
@@ -21,6 +21,7 @@
   <include package=".datamanager" />
   <include package=".restapi" />
   <include package=".captcha" />
+  <include package=".adapters" />
 
   <include file="permissions.zcml" />
   <include file="upgrades.zcml" />
