+++ src/collective/volto/formsupport/restapi/services/submit_form/post.py
@@ -1,12 +1,9 @@
+import codecs
+import logging
+import os
+import re
+
 from bs4 import BeautifulSoup
-from collective.volto.formsupport import _
-from collective.volto.formsupport.interfaces import ICaptchaSupport
-from collective.volto.formsupport.interfaces import IFormDataStore
-from collective.volto.formsupport.interfaces import IPostEvent
-from collective.volto.formsupport.utils import get_blocks
-from collective.volto.formsupport.events import FormSubmittedEvent
-from collective.volto.otp.utils import validate_email_token
-from copy import deepcopy
 from datetime import datetime
 from email import policy
 from email.message import EmailMessage
@@ -20,9 +17,7 @@ except ImportError:
 
 from plone.protect.interfaces import IDisableCSRFProtection
 from plone.registry.interfaces import IRegistry
-from plone.restapi.deserializer import json_body
 from plone.restapi.services import Service
-from plone.schema.email import _isemail
 from xml.etree.ElementTree import Element
 from xml.etree.ElementTree import ElementTree
 from xml.etree.ElementTree import SubElement
@@ -34,11 +29,11 @@ from zope.i18n import translate
 from zope.interface import alsoProvides
 from zope.interface import implementer
 
-import codecs
-import logging
-import math
-import os
-import re
+from collective.volto.formsupport import _
+from collective.volto.formsupport.interfaces import IFormDataStore
+from collective.volto.formsupport.interfaces import IPostEvent
+from collective.volto.formsupport.interfaces import IFormData
+from collective.volto.formsupport.events import FormSubmittedEvent
 
 
 logger = logging.getLogger(__name__)
@@ -57,14 +52,12 @@ class SubmitPost(Service):
         super().__init__(context, request)
 
         self.block = {}
-        self.form_data = self.cleanup_data()
+        self.form_data = self.get_form_data()
         self.block_id = self.form_data.get("block_id", "")
         if self.block_id:
             self.block = self.get_block_data(block_id=self.block_id)
 
     def reply(self):
-        self.validate_form()
-
         store_action = self.block.get("store", False)
         send_action = self.block.get("send", [])
 
@@ -97,200 +90,8 @@ class SubmitPost(Service):
 
         return {"data": self.form_data.get("data", [])}
 
-    def cleanup_data(self):
-        """
-        Avoid XSS injections and other attacks.
-
-        - cleanup HTML with plone transform
-        - remove from data, fields not defined in form schema
-        """
-        form_data = json_body(self.request)
-        fixed_fields = []
-        transforms = api.portal.get_tool(name="portal_transforms")
-
-        block = self.get_block_data(block_id=form_data.get("block_id", ""))
-        block_fields = [x.get("field_id", "") for x in block.get("subblocks", [])]
-
-        for form_field in form_data.get("data", []):
-            if form_field.get("field_id", "") not in block_fields:
-                # unknown field, skip it
-                continue
-            new_field = deepcopy(form_field)
-            value = new_field.get("value", "")
-            if isinstance(value, str):
-                stream = transforms.convertTo("text/plain", value, mimetype="text/html")
-                new_field["value"] = stream.getData().strip()
-            fixed_fields.append(new_field)
-        form_data["data"] = fixed_fields
-        return form_data
-
-    def validate_form(self):
-        """
-        check all required fields and parameters
-        """
-        if not self.block_id:
-            raise BadRequest(
-                translate(
-                    _("missing_blockid_label", default="Missing block_id"),
-                    context=self.request,
-                )
-            )
-        if not self.block:
-            raise BadRequest(
-                translate(
-                    _(
-                        "block_form_not_found_label",
-                        default='Block with @type "form" and id "$block" not found in this context: $context',
-                        mapping={
-                            "block": self.block_id,
-                            "context": self.context.absolute_url(),
-                        },
-                    ),
-                    context=self.request,
-                ),
-            )
-
-        if not self.block.get("store", False) and not self.block.get("send", []):
-            raise BadRequest(
-                translate(
-                    _(
-                        "missing_action",
-                        default='You need to set at least one form action between "send" and "store".',  # noqa
-                    ),
-                    context=self.request,
-                )
-            )
-
-        if not self.form_data.get("data", []):
-            raise BadRequest(
-                translate(
-                    _(
-                        "empty_form_data",
-                        default="Empty form data.",
-                    ),
-                    context=self.request,
-                )
-            )
-
-        self.validate_attachments()
-        if self.block.get("captcha", False):
-            getMultiAdapter(
-                (self.context, self.request),
-                ICaptchaSupport,
-                name=self.block["captcha"],
-            ).verify(self.form_data.get("captcha"))
-
-        self.validate_email_fields()
-        self.validate_bcc()
-
-    def validate_email_fields(self):
-        email_fields = [
-            x.get("field_id", "")
-            for x in self.block.get("subblocks", [])
-            if x.get("field_type", "") == "from"
-        ]
-        for form_field in self.form_data.get("data", []):
-            if form_field.get("field_id", "") not in email_fields:
-                continue
-            if _isemail(form_field.get("value", "")) is None:
-                raise BadRequest(
-                    translate(
-                        _(
-                            "wrong_email",
-                            default='Email not valid in "${field}" field.',
-                            mapping={
-                                "field": form_field.get("label", ""),
-                            },
-                        ),
-                        context=self.request,
-                    )
-                )
-
-    def validate_attachments(self):
-        attachments_limit = os.environ.get("FORM_ATTACHMENTS_LIMIT", "")
-        if not attachments_limit:
-            return
-        attachments = self.form_data.get("attachments", {})
-        attachments_len = 0
-        for attachment in attachments.values():
-            data = attachment.get("data", "")
-            attachments_len += (len(data) * 3) / 4 - data.count("=", -2)
-        if attachments_len > float(attachments_limit) * pow(1024, 2):
-            size_name = ("B", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB")
-            i = int(math.floor(math.log(attachments_len, 1024)))
-            p = math.pow(1024, i)
-            s = round(attachments_len / p, 2)
-            uploaded_str = f"{s} {size_name[i]}"
-            raise BadRequest(
-                translate(
-                    _(
-                        "attachments_too_big",
-                        default="Attachments too big. You uploaded ${uploaded_str},"
-                        " but limit is ${max} MB. Try to compress files.",
-                        mapping={
-                            "max": attachments_limit,
-                            "uploaded_str": uploaded_str,
-                        },
-                    ),
-                    context=self.request,
-                )
-            )
-
-    def validate_bcc(self):
-        """
-        If otp validation is enabled, check if is valid
-        """
-        bcc_fields = []
-        email_otp_verification = self.block.get("email_otp_verification", False)
-        block_id = self.form_data.get("block_id", "")
-        for field in self.block.get("subblocks", []):
-            if field.get("use_as_bcc", False):
-                field_id = field.get("field_id", "")
-                if field_id not in bcc_fields:
-                    bcc_fields.append(field_id)
-        if not bcc_fields:
-            return
-        if not email_otp_verification:
-            return
-        for data in self.form_data.get("data", []):
-            value = data.get("value", "")
-            if not value:
-                continue
-            if data.get("field_id", "") not in bcc_fields:
-                continue
-            otp = data.get("otp", "")
-            if not otp:
-                raise BadRequest(
-                    api.portal.translate(
-                        _(
-                            "otp_validation_missing_value",
-                            default="Missing OTP value. Unable to submit the form.",
-                        )
-                    )
-                )
-            if not validate_email_token(block_id, value, otp):
-                raise BadRequest(
-                    api.portal.translate(
-                        _(
-                            "otp_validation_wrong_value",
-                            default="${email}'s OTP is wrong",
-                            mapping={"email": data["value"]},
-                        )
-                    )
-                )
-
-    def get_block_data(self, block_id):
-        blocks = get_blocks(self.context)
-        if not blocks:
-            return {}
-        for id, block in blocks.items():
-            if id != block_id:
-                continue
-            block_type = block.get("@type", "")
-            if block_type != "form":
-                continue
-            return block
-        return {}
+    def get_form_data(self):
+        return getMultiAdapter((self.context, self.request), IFormData)()
 
     def get_reply_to(self):
         """This method retrieves the correct field to be used as 'reply to'.
